[DownloadWidget] Fix wrong file sizes
Due to a bug, files above 2GB had the wrong file size. That is because
an integer (32bit) was used to store the file size in Byte.
Furthermore MediaElch showed file sizes in MiB, GiB, etc. even though
MB, GB, etc. were displayed.

+++ src/ui/downloads/DownloadsWidget.h
@@ -21,7 +21,9 @@ class DownloadsWidget : public QWidget
     {
         QString baseName;
         QStringList files;
-        int64_t size;
+        /// Size in Bytes of this package.
+        /// Not an int to allow sizes >4GB on 32bit systems
+        double size;
     };
 
     struct Import
@@ -29,7 +31,9 @@ class DownloadsWidget : public QWidget
         QString baseName;
         QStringList files;
         QStringList extraFiles;
-        int64_t size;
+        /// Size in Bytes of this import.
+        /// Not an int to allow sizes >4GB on 32bit systems
+        double size;
     };
 
 public:
